Rename terraform_state_parser.py -> terraform_state.py
